Inject user's default organization if necessary
This PR adds a mutating policy which sets a namespace's
appuio.io/organization label to the user's default organization (sourced
from annotation appuio.io/default-organization on the User object) if
the namespace which is being created doesn't have the label.

This is implemented in the same cluster policy as the label checking,
which works as expected, as mutations are executed before validations.

Currently the annotations on User objects providing the information
about default organizations is generated semi-manually by a Go program
which extracts the user's default organization from Keycloak and writes
it to the User object.