Add organization label to projects created with `oc new-project`
Add a mutate policy which adds the requesting user's default
organization to the project based on the value of annotation
`openshift.io/requester` on the Project resource created by the OCP
project template.

Refactor the code a bit since this mutate policy is mostly identical to
the one which handles setting a default organization when users create
projects with `kubectl create ns`.