Kyverno policy renderer: extract policy types from parsed policies
The policy renderer doesn't actually check what the policy does to
determine if it's a "mutate", "generate", or "validate" policy, it just
does a string match on the policy YAML for each of those words and
returns the first one it finds as the policy type.

Previously the code searched "generate", "mutate", "validate" without
any fencing.  Since our excludes contain the string "validate" a lot of
policies were mislabelled.

This commit changes the implementation to instead go through the rules
on the parsed `kyvernov1.ClusterPolicy` object and identify all rule
types present in the policy.