Adjust default limitrange requests ratio
The component uses a default memory-per-core ratio of 4Gi. This commit
adjusts the default limitrange requests to match that ratio.

Note that 25m / 100Mi doesn't exactly match the 4Gi/core ratio, even
though the values appear to have a ratio of 1:4.

Instead of adjusting the memory requests to 102.4Mi to match the default
ratio exactly, we go with 10m / 100Mi for the default requests.

This should be a good default, since most applications really don't need
a lot of CPU when they're idle, and we leave the default limit at 600m /
768 Mi, so standard bursty workloads which don't have to serve hundreds
of queries per second should be completely fine with the new default
requests.