Add support for overriding storageclass-specific quotas
The storageclass-specific quotas use keys which are too long to work
with the generic annotation-based system for customizing the quotas in
single namespaces.

This commit introduces a special annotation
`resourcequota.appuio.io/<quota-name>.storageclasses` whose value is
parsed as a JSON object. If there's a key which matches an existing
`<storageclass>.storageclass.k8s.io` quota, the policy will use the
value of that key as the final value for that quota instead of the
default defined in the component defaults.